// src/types/health.ts
export enum UnderlyingHealthCondition {
  Diabetes = 'Diabetes',
  Hypertension = 'Hypertension',
  HeartDisease = 'Heart Disease',
  Cancer = 'Cancer',
  Thalassemia = 'Thalassemia',
  Hemophilia = 'Hemophilia',
  Epilepsy = 'Epilepsy',
  ActivePulmonaryTuberculosis = 'Active pulmonary tuberculosis',
  SevereAnemia = 'Severe anemia',
  SevereNeurologicalDisorder = 'Severe neurological disorder',
  HIV = 'HIV/AIDS',
  HepatitisBorC = 'Hepatitis B or C',
  None = 'None'
}